ai-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
pdfplumber==0.11.4
python-dotenv==1.0.1
python-multipart==0.0.20
google-genai==1.12.1
